pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smser"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "pyserial >= 3.5",
    "smspdudecoder >= 2.1.0",
    "pyTelegramBotAPI >= 4.22.1",
    "pydantic >= 2.9.1",
    "pyyaml >= 6.0.2",
]

[tool.setuptools.packages.find]
where = ["."]
